Hitfinder threshold can be a float

+++ strax/dtypes.py
@@ -91,7 +91,7 @@ hit_dtype = interval_dtype + [
     (('Internal (temporary) index of fragment in which hit was found',
         'record_i'), np.int32),
     (('ADC threshold applied in order to find hits',
-        'threshold'), np.int32),
+        'threshold'), np.float32),
     (('Maximum amplitude above baseline [ADC counts]',
         'height'), np.float32),
 ]
